# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "summary": "Signup",
        "operationId": "signup_api_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "summary": "Signin",
        "operationId": "signin_api_auth_signin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/reset-password": {
      "post": {
        "summary": "Reset Password",
        "operationId": "reset_password_api_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/update-password": {
      "post": {
        "summary": "Update Password",
        "operationId": "update_password_api_auth_update_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart/add": {
      "post": {
        "summary": "Add To Cart",
        "description": "Добавить товар в корзину",
        "operationId": "add_to_cart_api_cart_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItemInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/{item_id}": {
      "patch": {
        "summary": "Update Cart Item",
        "description": "Изменить количество товара в корзине",
        "operationId": "update_cart_item_api_cart__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItemInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove From Cart",
        "description": "Удалить товар из корзины",
        "operationId": "remove_from_cart_api_cart__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/merge": {
      "post": {
        "summary": "Merge Cart",
        "operationId": "merge_cart_api_cart_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CartItemInDB"
                  },
                  "type": "array",
                  "title": "Response Merge Cart Api Cart Merge Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CartItemAdd": {
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "productId"
        ],
        "title": "CartItemAdd"
      },
      "CartItemInDB": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "priceSnapshot": {
            "type": "number",
            "title": "Pricesnapshot"
          }
        },
        "type": "object",
        "required": [
          "id",
          "productId",
          "userId",
          "quantity",
          "priceSnapshot"
        ],
        "title": "CartItemInDB"
      },
      "CartItemUpdate": {
        "properties": {
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "quantity"
        ],
        "title": "CartItemUpdate"
      },
      "CartMergeRequest": {
        "properties": {
          "localCart": {
            "items": {
              "$ref": "#/components/schemas/LocalCartItem"
            },
            "type": "array",
            "title": "Localcart"
          }
        },
        "type": "object",
        "required": [
          "localCart"
        ],
        "title": "CartMergeRequest"
      },
...
      "CustomUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "fullName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname"
          },
          "isAdmin": {
            "type": "boolean",
            "title": "Isadmin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "isAdmin"
        ],
        "title": "CustomUser"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocalCartItem": {
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "priceSnapshot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Pricesnapshot"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "quantity",
          "priceSnapshot"
        ],
        "title": "LocalCartItem"
      },
...
      "ResetPasswordSchema": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ResetPasswordSchema"
      },
      "SignInSchema": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "SignInSchema"
      },
      "SignUpSchema": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6,
            "title": "Password"
          },
          "confirmPassword": {
            "type": "string",
            "title": "Confirmpassword"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "confirmPassword"
        ],
        "title": "SignUpSchema"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "UpdatePasswordSchema": {
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "UpdatePasswordSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ]
}